+++ Scripts/tsan-stress.sh
@@ -58,6 +58,7 @@ SCENARIOS=(
   "349-ocaf-driver-reentrancy/occt_349_barrier.cpp|8 50 @SCRATCH"
   "353-cdm-metadata-lookup-table/occt_353_barrier.cpp|8 50 @SCRATCH"
   "371-getapplication-singleton-elimination/occt_371_private_app.cpp|8 50 @SCRATCH"
+  "374-resource-manager-storage-schema-race/occt_374_stress.cpp|8 50 @SCRATCH"
 )
 
 MACOS_SDK=$(xcrun --sdk macosx --show-sdk-path)